Code improvements + fix for API response decoding
